requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langgraph
openai
pydantic
typing-extensions
python-dotenv
python-dateutil
tenacity
beautifulsoup4
ipdb
requests
duckduckgo-search
ddgs
pandas
yfinance
stockstats
tqdm
rank_bm25
bs4
langchain_openai
